third_party/WebKit/LayoutTests/storage/indexeddb/index-getall.html

<!DOCTYPE html>
<title>IndexedDB: Test IDBIndex.getAll.</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>

function doSetup(dbName, dbVersion, onsuccess) {
  var delete_request = indexedDB.deleteDatabase(dbName);
  delete_request.onerror = function() {
    assert_unreached('deleteDatabase should not fail');
  };
  delete_request.onsuccess = function(e) {
    var req = indexedDB.open(dbName, dbVersion);
    req.onsuccess = onsuccess;
    req.onerror = function() {
      assert_unreached('open should not fail');
    };
    req.onupgradeneeded = function(evt) {
      var connection = evt.target.result;
      var alphabet = 'abcdefghijklmnopqrstuvwxyz'.split('');

      var store = connection.createObjectStore('generated',
            {autoIncrement: true, keyPath: 'id'});
      var index = store.createIndex('test_idx', 'upper');
      alphabet.forEach(function(letter) {
        store.put({ch: letter, upper: letter.toUpperCase()});
      });

      store = connection.createObjectStore('out-of-line', null);
      index = store.createIndex('test_idx', 'upper');
      alphabet.forEach(function(letter) {
        store.put({ch: letter, upper: letter.toUpperCase()}, letter);
      });

      store = connection.createObjectStore('out-of-line-not-unique', null);
      index = store.createIndex('test_idx', 'half');
      alphabet.forEach(function(letter) {
        if (letter <= 'm')
          store.put({ch: letter, half: 'first'}, letter);
        else
          store.put({ch: letter, half: 'second'}, letter);
      });

      store = connection.createObjectStore('out-of-line-multi', null);
      index = store.createIndex('test_idx', 'attribs', {multiEntry: true});
      alphabet.forEach(function(letter) {
        attrs = [];
        if (['a', 'e', 'i', 'o', 'u'].indexOf(letter) != -1)
          attrs.push('vowel');
        else
          attrs.push('consonant');
        if (letter == 'a')
          attrs.push('first');
        if (letter == 'z')
          attrs.push('last');
        store.put({ch: letter, attribs: attrs}, letter);
      });

      store = connection.createObjectStore('empty', null);
      index = store.createIndex('test_idx', 'upper');
    };
  };
}

function createGetAllRequest(t, storeName, connection, range, maxCount) {
    var transaction = connection.transaction(storeName, 'readonly');
    var store = transaction.objectStore(storeName);
    var index = store.index('test_idx');
    // TODO(cmumford): Simplify once crbug.com/335871 is fixed.
    var req = maxCount !== undefined ? index.getAll(range, maxCount) :
              range !== undefined ? index.getAll(range) :  index.getAll();
    req.onerror = t.unreached_func('getAll request should succeed');
    return req;
}

doSetup(location.pathname + '-IDBIndex.getAll', 1, function(evt) {
    var connection = evt.target.result;
    async_test(function(t) {
      var req = createGetAllRequest(t, 'out-of-line', connection, 'C');
      req.onsuccess = t.step_func(function(evt) {
          var data = evt.target.result;
          assert_class_string(data, 'Array', 'result should be an array');
          assert_equals(data.length, 1);
          assert_object_equals(data[0], {ch: 'c', upper: 'C'});
          t.done();
      });
    }, 'Single item get');

    async_test(function(t) {
      var req = createGetAllRequest(t, 'empty', connection);
      req.onsuccess = t.step_func(function(evt) {
          assert_array_equals(evt.target.result, [],
              'getAll() on empty object store should return an empty array');
          t.done();
      });
    }, 'Empty object store');

    async_test(function(t) {
      var req = createGetAllRequest(t, 'out-of-line', connection);
      req.onsuccess = t.step_func(function(evt) {
          var data = evt.target.result;
          assert_class_string(data, 'Array', 'result should be an array');
          assert_equals(data.length, 26);
          assert_object_equals(data[0], {ch: 'a', upper: 'A'});
          assert_object_equals(data[25], {ch: 'z', upper: 'Z'});
          t.done();
      });
    }, 'Get all keys');

    async_test(function(t) {
      var req = createGetAllRequest(t, 'out-of-line', connection, undefined,
                                    10);
      req.onsuccess = t.step_func(function(evt) {
          var data = evt.target.result;
          assert_class_string(data, 'Array', 'result should be an array');
          assert_equals(data.length, 10);
          assert_object_equals(data[0], {ch: 'a', upper: 'A'});
          assert_object_equals(data[9], {ch: 'j', upper: 'J'});
          t.done();
      });
    }, 'maxCount=10');

    async_test(function(t) {
      var req = createGetAllRequest(t, 'out-of-line', connection,
                                    IDBKeyRange.bound('G', 'M'));
      req.onsuccess = t.step_func(function(evt) {
          var data = evt.target.result;
          assert_equals(data.length, 7);
          assert_object_equals(data[0], {ch: 'g', upper: 'G'});
          assert_object_equals(data[6], {ch: 'm', upper: 'M'});
          t.done();
      });
    }, 'Get bound range');

    async_test(function(t) {
      var req = createGetAllRequest(t, 'out-of-line', connection,
                                    IDBKeyRange.bound('G', 'M'), 3);
      req.onsuccess = t.step_func(function(evt) {
          var data = evt.target.result;
          assert_class_string(data, 'Array', 'result should be an array');
          assert_equals(data.length, 3);
          assert_object_equals(data[0], {ch: 'g', upper: 'G'});
          assert_object_equals(data[2], {ch: 'i', upper: 'I'});
          t.done();
      });
    }, 'Get bound range with maxCount');

    async_test(function(t) {
      var req = createGetAllRequest(t, 'out-of-line', connection,
          IDBKeyRange.bound('G', 'K', false, true));
      req.onsuccess = t.step_func(function(evt) {
          var data = evt.target.result;
          assert_class_string(data, 'Array', 'result should be an array');
          assert_equals(data.length, 4);
          assert_object_equals(data[0], {ch: 'g', upper: 'G'});
          assert_object_equals(data[1], {ch: 'h', upper: 'H'});
          assert_object_equals(data[2], {ch: 'i', upper: 'I'});
          assert_object_equals(data[3], {ch: 'j', upper: 'J'});
          t.done();
      });
    }, 'Get upper excluded');

    async_test(function(t) {
      var req = createGetAllRequest(t, 'out-of-line', connection,
          IDBKeyRange.bound('G', 'K', true, false));
      req.onsuccess = t.step_func(function(evt) {
          var data = evt.target.result;
          assert_class_string(data, 'Array', 'result should be an array');
          assert_equals(data.length, 4);
          assert_object_equals(data[0], {ch: 'h', upper: 'H'});
          assert_object_equals(data[1], {ch: 'i', upper: 'I'});
          assert_object_equals(data[2], {ch: 'j', upper: 'J'});
          assert_object_equals(data[3], {ch: 'k', upper: 'K'});
          t.done();
      });
    }, 'Get lower excluded');

    async_test(function(t) {
      var req = createGetAllRequest(t, 'generated',
          connection, IDBKeyRange.bound(4, 15), 3);
      req.onsuccess = t.step_func(function(evt) {
          var data = evt.target.result;
          assert_true(Array.isArray(data));
          assert_equals(data.length, 0);
          t.done();
      });
    }, 'Get bound range (generated) with maxCount');

    async_test(function(t) {
      var req = createGetAllRequest(t, 'out-of-line',
          connection, "Doesn't exist");
      req.onsuccess = t.step_func(function(evt) {
          assert_array_equals(evt.target.result, [],
              'getAll() using a nonexistent key should return an empty array');
          t.done();
      req.onerror = t.unreached_func('getAll request should succeed');
      });
    }, 'Non existent key');

    async_test(function(t) {
      var transaction = connection.transaction('out-of-line', 'readonly');
      var store = transaction.objectStore('out-of-line');
      var index = store.index('test_idx');
      assert_throws(new TypeError(), function () {
        index.getAll(undefined, 0);
      }, 'getAll() with maxCount=0 should throw TypeError');
      t.done();
    }, 'maxCount=0');

    async_test(function(t) {
      var req = createGetAllRequest(t, 'out-of-line-not-unique', connection,
                                    'first');
      req.onsuccess = t.step_func(function(evt) {
          var data = evt.target.result;
          assert_class_string(data, 'Array', 'result should be an array');
          assert_equals(data.length, 13);
          assert_object_equals(data[0], {ch: 'a', half: 'first'});
          assert_object_equals(data[12], {ch: 'm', half: 'first'});
          t.done();
      });
    }, 'Retrieve multiEntry key');

    async_test(function(t) {
      var req = createGetAllRequest(t, 'out-of-line-multi', connection,
                                    'vowel');
      req.onsuccess = t.step_func(function(evt) {
          var data = evt.target.result;
          assert_class_string(data, 'Array', 'result should be an array');
          assert_equals(data.length, 5);
          assert_object_equals(data[0], {ch: 'a', attribs: ['vowel', 'first']});
          assert_object_equals(data[1], {ch: 'e', attribs: ['vowel']});
          assert_object_equals(data[2], {ch: 'i', attribs: ['vowel']});
          assert_object_equals(data[3], {ch: 'o', attribs: ['vowel']});
          assert_object_equals(data[4], {ch: 'u', attribs: ['vowel']});
          t.done();
      });
    }, 'Retrieve one key multiple values');
});

</script>
